pyproject.toml
[project]
name = "gradio-sso-auth-descope"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "descope>=1.7.1",
    "flask>=3.1.0",
    "gradio>=5.9.1",
    "python-dotenv>=1.0.1",
]
